refactor(android): auth
applied new layout
