+++ apps/android/app/src/main/java/at/aau/serg/android/ui/theme/AuthColors.kt
@@ -1,0 +1,17 @@
+package at.aau.serg.android.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+data class AuthColors(
+    val background: Color,
+    val card: Color,
+    val input: Color,
+    val primaryText: Color,
+    val secondaryText: Color,
+    val cardBlue: Color,
+    val borderBlue: Color,
+    val cardPurple: Color,
+    val borderPurple: Color,
+    val cardGreen: Color,
+    val borderGreen: Color,
+)
